move external namespaces to a separate header file
